requirements.txt
appdirs==1.4.3
docker==2.2.1
docker-pycreds==0.2.1
packaging==16.8
pyparsing==2.2.0
PyYAML==3.12
requests==2.13.0
six==1.10.0
websocket-client==0.40.0
